Move components to separate folders
