+++ frontend/smart-kickers-game/src/components/Game/GameStatistics/GameStatistics.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './GameStatistics.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { Button } from '../Button/Button.js';
+import { Button } from '../../Button/Button.js';
 
 function GameStatistics({ finalScores, setIsStatisticsDisplayed }) {
   return (
@@ -11,12 +11,12 @@ function GameStatistics({ finalScores, setIsStatisticsDisplayed }) {
       </h2>
       <div className="table-with-stats">
         <div className="table-item">
-          <FontAwesomeIcon className="blueTeamIcon" icon="fa-person" />
+          <FontAwesomeIcon className="blue-team-icon" icon="fa-person" />
           Blue
         </div>
         <div className="table-item"></div>
         <div className="table-item">
-          <FontAwesomeIcon className="whiteTeamIcon" icon="fa-person" />
+          <FontAwesomeIcon className="white-team-icon" icon="fa-person" />
           White
         </div>
         <div className="table-item">{finalScores.blue}</div>
